10l correct type for =a and =d constraints under x86-32
git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@30757 b3059339-0415-0410-9bf9-f77b7e298cf2

+++ cpuinfo.c
@@ -77,10 +77,10 @@ cpuid(int func) {
 static int64_t
 rdtsc(void)
 {
-  uint64_t hi, lo;
+  uint32_t hi, lo;
 #define RDTSC   ".byte 0x0f, 0x31; "
   __asm__ volatile (RDTSC : "=a"(lo), "=d"(hi) : );
-  return hi << 32 | lo;
+  return (uint64_t) hi << 32 | lo;
 }
 
 static const char*
